iop: use the new RemoveRange function
IOP is a bit tricky because block don't stop at page boundary.
So previous page is checked too, and again until a page is left
untouched

It is likely slower that previous method. Hopefully function
isn't enough called to impact the performance.

However it reduces code complexity for the future